Try to fix too long token bug in login

+++ src/ServerSide/src/main/java/vn/zalopay/ducnm8/da/ChatListDA.java
@@ -1,12 +1,12 @@
 package vn.zalopay.ducnm8.da;
 
-import vn.zalopay.ducnm8.model.Chat;
 import io.vertx.core.Future;
+import vn.zalopay.ducnm8.model.Chat;
 
 import java.util.List;
 
 public interface ChatListDA {
     Executable<Chat> insert(Chat chat);
 
-    Future<List<Chat>> listChatByMember(int UserSendID, int UserReceiveID);
+    Future<List<Chat>> listChatByMember(long UserSendID, long UserReceiveID);
 }
